requirements.txt
Flask==1.1.2
Flask-Caching==1.9.0
psycopg2-binary==2.8.6
